feat: build and serve natively on Windows with VS2022 and CUDA 13.3
The engine, CLI and HTTP server now build with MSVC 19.44 and CUDA 13.3, and the TP-2 server runs a
27B NVFP4 artifact at Linux parity on two RTX 5060 Ti cards (plain decode 32.9 vs 32.8 tok/s, MTP
--draft-tokens 2 56.9 vs 56.8 tok/s, media turns included).

Platform branches:
- cmake: Unix keeps pkg-config; Windows discovers FFmpeg and the libcurl import library from explicit
  prefixes, and MSVC needs /Zc:preprocessor (CUDA 13 CCCL), /utf-8 (bundled fmt), NOMINMAX and
  WIN32_LEAN_AND_MEAN
- artifact file_io: CreateFileW/ReadFile with OVERLAPPED positional reads and a platform handle type
- core/uint128.h: 128-bit accounting without __int128 (saturating add, shift, compare)
- core/host_process.h: process id, stderr tty and localtime/gmtime for the per-platform CRTs
- nvfp4 TMA kernels: a 64-byte-aligned tensor map cannot be passed by value under the MSVC parameter
  ABI (C2719), so Nvfp4TmaDescriptorStaging uploads descriptors into stream-ordered device memory
- media_acquire: winsock name resolution with a one-time WSAStartup, linked against ws2_32
- logging: console width from GetConsoleScreenBufferInfo; utf8proc builds with UTF8PROC_STATIC

Tooling and docs: tools/win_port (PowerShell vcvars/build/serve/spike plus the mapped-pinned spike
source and the ffmpeg/libcurl fetchers), docs/windows.md, and a build_r35.sh sync that carries the
cmake directory.

+++ src/runtime/engine/tp2_generation_core.cpp
@@ -970,7 +970,8 @@ GenerationResult TP2GenerationCore::execute(Request& request, OutputSink* sink,
             }
             Tensor window_logits   = ws_a.alloc(DType::BF16, {vocab, width});
             round_hidden           = ws_a.alloc(DType::BF16, {hidden, width});
-            Tensor window_drafts   = ws_a.alloc(DType::I32, {mtp_drafts_, 1});
+            Tensor window_drafts =
+                ws_a.alloc(DType::I32, {static_cast<std::int32_t>(mtp_drafts_), 1});
             Tensor target_tokens   = ws_a.alloc(DType::I32, {width, 1});
             Tensor current_extents = ws_a.alloc(DType::I32, {1});
             Tensor round_lengths   = ws_a.alloc(DType::I32, {1});
